% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arl.R
\name{arl_businesses}
\alias{arl_businesses}
\title{Get Arlington County business licenses}
\usage{
arl_businesses(top = 1000, skip = 0)
}
\arguments{
\item{top}{Number of records (default 1000)}

\item{skip}{Pagination offset (default 0)}
}
\value{
tibble with active business license data
}
\description{
Get Arlington County business licenses
}
