% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arl.R
\name{arl_sales}
\alias{arl_sales}
\title{Get Arlington County property sales history}
\usage{
arl_sales(top = 1000, skip = 0)
}
\arguments{
\item{top}{Number of records (default 1000)}

\item{skip}{Pagination offset (default 0)}
}
\value{
tibble with property sale records
}
\description{
Get Arlington County property sales history
}
