% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arl.R
\name{arl_service_requests}
\alias{arl_service_requests}
\title{Get Arlington County service requests}
\usage{
arl_service_requests(top = 1000, skip = 0)
}
\arguments{
\item{top}{Number of records (default 1000)}

\item{skip}{Pagination offset (default 0)}
}
\value{
tibble with 311 service request data
}
\description{
Get Arlington County service requests
}
