% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arl.R
\name{arl_trade_names}
\alias{arl_trade_names}
\title{Get Arlington County trade name registrations}
\usage{
arl_trade_names(top = 1000, skip = 0)
}
\arguments{
\item{top}{Number of records (default 1000)}

\item{skip}{Pagination offset (default 0)}
}
\value{
tibble with trade name records
}
\description{
Get Arlington County trade name registrations
}
